% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos_map.R
\name{add_perc}
\alias{add_perc}
\title{circos_map}
\usage{
add_perc(mapRes, cell_num_list)
}
\arguments{
\item{mapRes}{A dataframe of the output of function cluster_map_by_marker.}

\item{cell_num_list}{A list of vector of cell numbers for each group and each sample.}
}
\value{
A dataframe of the matching results with cell percentage column added.
}
\description{
Plot Circos plot for the matching results.
}
