% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermap.R
\name{blocks}
\alias{blocks}
\title{blocks}
\usage{
blocks()
}
\description{
This function creates a small example gene expression data set with 110 genes as
  rows and 40 samples as columns.
}
\examples{
\dontrun{
tmp = blocks() # Contains the 110x40-matrix X and 40-vectors y1,y2,y3
X = tmp$X
y1 = tmp$y1 # Additional information track for the columns
y2 = tmp$y2 # Additional information track for the columns
y3 = tmp$y3 # Additional information track for the columns
}
}
\keyword{example}
