% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos_map.R
\name{circos_map}
\alias{circos_map}
\title{circos_map}
\usage{
circos_map(mapRes, cell_num_list, output, color_cord = NULL,
  color_sample = NULL)
}
\arguments{
\item{mapRes}{A dataframe of the output of function cluster_map_by_marker.}

\item{cell_num_list}{A list of vector of cell numbers for each group and each sample.}

\item{output}{The output directory to save the plot.}

\item{color_cord}{A vector of colors for the cord of circos plot. DEFAULT is NULL. Pre defined internal color will be used.}

\item{color_sample}{A vector of colors for the sample sectors in the circos plot. DEFAULT is NULL. Pre defined internal color will be used.}
}
\value{
circos plot will be save.
}
\description{
Plot Circos plot for the matching results.
}
