% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master.R
\name{cluster_map}
\alias{cluster_map}
\title{cluster_map}
\usage{
cluster_map(marker_file_list, edge_cutoff = 0.1, output,
  cell_num_list = NULL, single_obj_list = NULL, comb_obj = NULL,
  comb_delim = "-", k = 5)
}
\arguments{
\item{marker_file_list}{A list of csv files with names. Each file is a marker gene table for a sample. The columns named as 'cluster' and 'gene' are required.}

\item{edge_cutoff}{The edge length cutoff to decide the sub-nodes to merge or not. DEFAULT is 0.1.}

\item{output}{The output directory to save the matching results.}

\item{cell_num_list}{A list of vector of cell numbers for each group and each sample.}

\item{single_obj_list}{A list of Seurat object for each sample, with the same list names as the list names of marker_file_list.}

\item{comb_obj}{A Seurat object for the combined sample. Cells in different samples are labelled by the sample names with the comb_delim. The sample names should be the same as the list names of marker_file_list.}

\item{comb_delim}{The delimiter used in the cell names in the combined object to connect sample name and cell name in individual sample. DEFAULT is '-'.}

\item{k}{K-nearest neighbours used to calculate distance. DEFAULT is 5.}
}
\value{
A dataframe of the matching results. Heatmap of marker genes, the corresponding dendrogram, circos plot and recolored t-SNE plots will be saved into files.
}
\description{
A master function to perform the full workflow of ClusterMap.
}
