% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_map.R
\name{cluster_map_by_marker}
\alias{cluster_map_by_marker}
\title{cluster_map_by_marker}
\usage{
cluster_map_by_marker(marker_file_list, cutoff = 0.1, output)
}
\arguments{
\item{marker_file_list}{A list of csv files. Each file is a marker gene table for a sample. The columns named as 'cluster' and 'gene' are required.}

\item{cutoff}{The edge length cutoff to decide the sub-nodes to merge or not. DEFAULT is 0.1.}

\item{output}{The output directory to save the matching results.}
}
\value{
A dataframe of the matching results. Heatmap of marker genes and the dendrogram will be saved into files.
}
\description{
Match groups by marker genes and decompose into new groups by purity cut.
}
