% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermap.R
\docType{package}
\name{clustermap}
\alias{clustermap}
\alias{clustermap-package}
\title{clustermap}
\description{
Clustermap is an R package for heatmap and hierarchical clustering visualization
}
\author{
Ole Christian Lingjærde & Chloé B. Steen
}
