% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermap.R
\name{col2num}
\alias{col2num}
\title{col2num}
\usage{
col2num(X, clust = "col")
}
\arguments{
\item{X}{matrix}

\item{clust}{String. Direction for clustering, default is "col" (column-wise)}
}
\description{
Function to transform raster to numerical matrix.
}
