% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermap.R
\name{color.cont}
\alias{color.cont}
\title{color.cont}
\usage{
color.cont(x, label, color, na.color)
}
\arguments{
\item{x}{numeric matrix}

\item{label}{String}

\item{color}{String. Color string separated by "-".}

\item{na.color}{String. Set colors to missing data. Default is "grey".}
}
\description{
Sets colors for continuous variables. Called by set.color().
}
