% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermap.R
\name{color.surv}
\alias{color.surv}
\title{color.surv}
\usage{
color.surv(x, label, color, na.color, clust)
}
\arguments{
\item{x}{numeric matrix}

\item{label}{String}

\item{color}{String. Color string separated by "-".}

\item{na.color}{String. Set colors to missing data. Default is "grey".}

\item{clust}{String. Direction for clustering.}
}
\description{
Sets colors for survival variables. Called by set.color().
}
