% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermap.R
\name{compute.hmap}
\alias{compute.hmap}
\title{compute.hmap}
\usage{
compute.hmap(X, colorscale, xmax, col.na)
}
\arguments{
\item{X}{matrix}

\item{colorscale}{String. Color string separated by "-".}

\item{xmax}{Integer Default is set to maximum value in input matrix X}

\item{col.na}{String. Specifies color of missing data. Default set to "grey".}
}
\description{
Helper function to be used by draw.hmap
}
