% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermap.R
\name{draw.fork.left}
\alias{draw.fork.left}
\title{draw.for.left}
\usage{
draw.fork.left(i, clust, trim, groups, delta, delta2, sep, lwd)
}
\arguments{
\item{i}{argument}

\item{clust}{argument}

\item{trim}{argument}

\item{groups}{argument}

\item{delta}{argument}

\item{delta2}{argument}

\item{sep}{argument}

\item{lwd}{Float. Line width.}
}
\description{
Plots on the left side of the heatmap. Function called by draw.tree()
}
