% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermap.R
\name{draw.hmap.key}
\alias{draw.hmap.key}
\title{draw.hmap.key}
\usage{
draw.hmap.key()
}
\description{
The function draw.hmap.key() creates a color key, i.e. a graphical overview
of what numerical values the different colors found in the heatmap correspond to.
}
