% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermap.R
\name{draw.init}
\alias{draw.init}
\title{draw.init()}
\usage{
draw.init(tree = c(), text = c(), cbar = c(), ckey = T, legend = F,
  inner = c(), outer = c())
}
\arguments{
\item{tree}{Integer. Adds a dendogram to the plot on the sides specified by the argument. Default is c(2,3),
which adds a dendogram on the left and top sides of the draw.}

\item{text}{Integer. Adds text on specified sides.}

\item{cbar}{Integer. Adds color bar(s) on specified sides.}

\item{ckey}{Boolean. Adds color bar key(s) (legends). Default is TRUE.}

\item{legend}{Boolean. Default is FALSE. If set to TRUE, color bar keys are plotted on same plot as heatmap.
Otherwise, it is plotted on a separate page.}

\item{inner}{Integer vector of 4 elements. Adds (or subtracts) a specific amount of space in the inner margins
on a specified side of the heatmap.}

\item{outer}{Integer vector of 4 elements. Adds (or subtracts) a specific amount of space in the outer margins
on a specified side of the heatmap.}
}
\description{
Initiates a clustermap draw. The purpose is to make
room on the four sides of the heatmap for additional elements such
as cluster dendrograms, color bars with additional info, and labels.
Supply as arguments the sides where you want such elements, with the
coding side=1 (below), side=2 (left), side=3 (top), side=4 (right).
}
\examples{
\dontrun{
draw.init() # Make no room for additional elements
draw.init(tree=c()) # Dendrogram on the left and on the top
draw.init(tree=3, text=4)  # Dendrogram on the top, labels on the right
}
}
