% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermap.R
\name{draw.silhouette}
\alias{draw.silhouette}
\title{draw.silhouette}
\usage{
draw.silhouette(clust = "col", ...)
}
\arguments{
\item{clust}{String. Sets direction of clustering, default is "col" (column-wise)}

\item{...}{list}
}
\description{
Plot Silhouettes for cluster diagnostics.
}
