% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermap.R
\name{draw.tree}
\alias{draw.tree}
\title{draw.tree}
\usage{
draw.tree(groups, side = 3, sep = 0.03, lwd = 0.5, trim = T,
  colors = NA)
}
\arguments{
\item{groups}{Identified groups}

\item{side}{Integer. Side onto which the tree is plotted on the heatmap. Default is 3 (top)}

\item{sep}{Float.}

\item{lwd}{Float. Line width of tree branches.}

\item{trim}{Boolean. Default set to TRUE.}

\item{colors}{String. Default set to NA. Colors the clusters.}
}
\description{
Adds a dendogram to heatmap.
}
