% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermap.R
\name{gap.binary}
\alias{gap.binary}
\title{gap.binary}
\usage{
gap.binary(X, linkage, B, K = 6)
}
\arguments{
\item{X}{matrix}

\item{linkage}{Linkage for clustering.}

\item{B}{integer.}

\item{K}{integer. Default set to 6.}
}
\description{
Identifies number of clusters using method "gap".
}
