% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermap.R
\name{gap.curve}
\alias{gap.curve}
\title{gap.curve()}
\usage{
gap.curve(X, clust = "col", distance = "euclidean", linkage = "complete",
  B = 100, K = 6)
}
\arguments{
\item{X}{matrix}

\item{clust}{String.}

\item{distance}{Default is set to "eucledian". Distance measure used for hierarchical clustering}

\item{linkage}{Default is "complete". Linkage for hierarchical clustering.}

\item{B}{integer. Default is 100.}

\item{K}{integer. Default is 6.}
}
\description{
gap.curve() function
}
