% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermap.R
\name{gap.optimal}
\alias{gap.optimal}
\title{gap.optimal()}
\usage{
gap.optimal(X, clust = "col", distance = "euclidean",
  linkage = "complete", B = 100)
}
\arguments{
\item{X}{matrix}

\item{clust}{String. Direction of clustering.}

\item{distance}{Default is set to "eucledian". Distance measure used for hierarchical clustering}

\item{linkage}{Default is "complete". Linkage for hierarchical clustering.}

\item{B}{integer. Default is 100.}
}
\description{
gap.optimal function
}
