% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermap.R
\name{get.order}
\alias{get.order}
\title{get.order}
\usage{
get.order(clust = "col")
}
\arguments{
\item{clust}{String. Default is "col". Specifies column-wise ("col") or row-rise clustering ("row").}
}
\description{
Retrieves order of samples after clustering.
}
