% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermap.R
\name{get.silhouette}
\alias{get.silhouette}
\title{plot.silhouette}
\usage{
get.silhouette(clust = "col", order = "tree")
}
\arguments{
\item{clust}{String. Sets direction of clustering, default is "col" (column-wise)}

\item{order}{String. Default is set to "tree"}
}
\description{
Get silhouettes
}
