% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermap.R
\name{get.subclust}
\alias{get.subclust}
\title{get.subclust}
\usage{
get.subclust(clust = "col", labels = c(), order = "tree")
}
\arguments{
\item{clust}{String. Retrieves cluster labels for column-wise clusters ("col", default) or
row-wise clusters ("row").}

\item{labels}{String vector. If you supply row/column labels using the argument 'labels' the function
returns a data frame with two columns, the first being the cluster labels and the second being the
supplied row/column labels.}

\item{order}{String. Get subclusters in same order as dendogram.}
}
\description{
Retrieves a vector of cluster labels after identification of subclusters with subclust().
}
