% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermap.R
\name{hcluster}
\alias{hcluster}
\title{hcluster}
\usage{
hcluster(X, clust = "col", distance = "euclidean", linkage = "complete",
  reorder = T, ncomp = NA)
}
\arguments{
\item{clust}{String. Perform clustering of rows or columns, can be by column: "col" (default) or by row: "row".}

\item{distance}{String. Distance measure for hierarchical clustering, can be"euclidean" (default), "pearson", "spearman".}

\item{linkage}{String. Method for clusterting, can be "complete" (default), "average", "single", "ward".}

\item{reorder}{Boolean. When TRUE, arranges the clustered items in optimized order.}

\item{ncomp}{Integer. Default NA. If set to integer k, only the first k principal components are used in clustering}
}
\description{
Performs hierarchical clustering of a data matrix X. To cluster both rows and columns,
call it twice with clust="col"and clust="row" respectively. The option reorder=TRUE leads to a nice
arrangement of the clusters from left to right to ensure neighbors are similar to each other.
}
