% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separability.R
\name{inna_dist}
\alias{inna_dist}
\title{inna_dist}
\usage{
inna_dist(x, k)
}
\arguments{
\item{x}{A distance matrix of cells in one sample to the cells in the same sample.}

\item{k}{K-nearest neighbours used to calculate distance.}
}
\value{
A single value of distance.
}
\description{
Calculate the inna-sample distance. Internal function called by separability.
}
