% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separability.R
\name{inter_dist}
\alias{inter_dist}
\title{inter_dist}
\usage{
inter_dist(x, k)
}
\arguments{
\item{x}{A distance matrix of cells in one sample to the cells in another sample.}

\item{k}{K-nearest neighbours used to calculate distance. DEFAULT is 5.}
}
\value{
A single value of distance.
}
\description{
Calculate the inter-sample distance. Internal function called by separability.
}
