% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_analysis.R
\name{make_comb_obj}
\alias{make_comb_obj}
\title{make_comb_obj}
\usage{
make_comb_obj(data_dirList, is.10X = TRUE, comb_delim = "-")
}
\arguments{
\item{data_dirList}{A list of directory with 10X genomics single cell data or full path of expression table.}

\item{is.10X}{The input data is 10X genomics format or not. DEFAULT is TRUE.}

\item{comb_delim}{The delimiter used in the cell names in the combined object to connect sample name and cell name. DEFAULT is '-'.}
}
\value{
A Seurat object.
}
\description{
A warper of Seurat function to generate Seurat object for combined sample from single sample.
}
