% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_analysis.R
\name{make_single_obj}
\alias{make_single_obj}
\title{make_single_obj}
\usage{
make_single_obj(data_dir, is.10X = TRUE, output)
}
\arguments{
\item{data_dir}{Directory with 10X genomics single cell data or full path of expression table.}

\item{is.10X}{The input data is 10X genomics format or not. DEFAULT is TRUE.}

\item{output}{The output directory to save the marker genes.}
}
\value{
A Seurat object.
}
\description{
A warper of Seurat function to generate Seurat object and marker genes for single sample.
}
