% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermap.R
\name{num2col}
\alias{num2col}
\title{num2col}
\usage{
num2col(Z, clust = "col")
}
\arguments{
\item{Z}{matrix}

\item{clust}{String. Default is "col". Specifies column-wise ("col") or row-rise clustering ("row").}
}
\description{
Function to transform numerical matrix to raster.
}
