% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermap.R
\name{panel.init}
\alias{panel.init}
\title{panel.init}
\usage{
panel.init(mfrow = c(2, 2))
}
\arguments{
\item{Integer}{Number of rows and columns for panel. Default "mfrow=c(2,2)".}
}
\description{
Initiates panel for plotting heatmaps.
}
