% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermap.R
\name{part.binary}
\alias{part.binary}
\title{part.binary}
\usage{
part.binary(X, linkage, B, minSize, maxLevel, K = 6)
}
\arguments{
\item{X}{matrix}

\item{linkage}{Default is "complete". Linkage for hierarchical clustering.}

\item{B}{integer. Default is 100.}

\item{minSize}{Integer Minimum size of cluster. Default set to 10.}

\item{maxLevel}{Integer. Maximum depth/level for clustering, default set to 3.}

\item{K}{Integer. Default is set to 6.}
}
\description{
Identifies number of clusters using method "part".
}
