% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermap.R
\name{part.euclidean}
\alias{part.euclidean}
\title{part.euclidean}
\usage{
part.euclidean(X, linkage, B, minSize, maxLevel, K = 6)
}
\arguments{
\item{X}{matrix}

\item{linkage}{Default is "complete". Linkage for hierarchical clustering.}

\item{B}{integer.}

\item{minSize}{Integer Minimum size of cluster.}

\item{maxLevel}{Integer. Maximum depth/level for clustering.}

\item{K}{integer. Default set to 6.}

\item{clust}{String. Direction of clustering.}
}
\description{
Identifies number of clusters using method "part" with Euclidean distance.
}
