% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermap.R
\name{part.new}
\alias{part.new}
\title{part.new}
\usage{
part.new(X, clust = "col", distance = "euclidean", linkage = "complete",
  B = 100, minSize = 10, maxLevel = 3)
}
\arguments{
\item{X}{matrix}

\item{clust}{String. Direction of clustering.}

\item{distance}{Default is set to "eucledian". Distance measure used for hierarchical clustering}

\item{linkage}{Default is "complete". Linkage for hierarchical clustering.}

\item{B}{integer. Default is 100.}

\item{minSize}{Integer Minimum size of cluster. Default set to 10.}

\item{maxLevel}{Integer. Maximum depth/level for clustering, default set to 3.}
}
\description{
Identifies number of clusters using method "part".
}
