% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermap.R
\name{part.spearman}
\alias{part.spearman}
\title{part.spearman}
\usage{
part.spearman(X, linkage, B, minSize, maxLevel, K = 6)
}
\arguments{
\item{X}{matrix}

\item{linkage}{Linkage for clustering.}

\item{B}{integer.}

\item{minSize}{Integer Minimum size of cluster.}

\item{maxLevel}{Integer. Maximum depth/level for clustering.}

\item{K}{integer. Default set to 6.}
}
\description{
Identifies number of clusters using method "part" with spearman correlation as distance measure.
}
