% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermap.R
\name{plot.cbar.key}
\alias{plot.cbar.key}
\title{plot.cbar.key}
\usage{
\method{plot}{cbar.key}(border = F, vsize = 0.1, hsize = 0.05,
  hsep = 0.05, vlsep = 0.1, vusep = 0.1, vsize.title = 0.08,
  cex.title = 1, font.title = 2, cex = 0.8, mfrow = NA)
}
\arguments{
\item{border}{Boolean. Default set to FALSE.}

\item{vsize}{float. Height of the key. Default set to 0.1}

\item{hsize}{float. Width of the key. Default set to 0.2}

\item{hsep}{float. Horizontal separator of the colorbar keys, default set to 0.05.}

\item{vlsep}{float. Vertical distance between lower part of the heatmap, and the bottom of the color bar.
Default is set to 0.1}

\item{vusep}{float. Vertical distance between upper part of the heatmap, and the bottom of the color bar.
Default is set to 0.1}

\item{vsize.title}{float. Space for adding color bar title. Default is 0.08}

\item{cex.title}{integer. Size of title font. Default is 1.}

\item{font.title}{integer. Font type for title, default is set to 2 (bold)}

\item{cex}{float. Text size for legend. Default is set to 0.8.}

\item{mfrow}{integer vector of length 2, default set to NA.}
}
\description{
Adds key for color bar.
}
