% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermap.R
\name{plot.cbar}
\alias{plot.cbar}
\title{plot.cbar}
\usage{
\method{plot}{cbar}(..., side = 3, labels = T, pvalue = F,
  pvalue.method = "chisq", cex = 0.8, sep.outer = 0.02, sep.inner = 0.2,
  border = NA, lwd = 0.5)
}
\arguments{
\item{...}{A list of set.color or color.cont objects}

\item{side}{Integer. Default is 3. Add the color bar to a specific side of the heatmap.}

\item{labels}{Boolean. Default is TRUE. Add labels to the colorbar.}

\item{pvalue}{Boolean. Default is FALSE. If TRUE, adds a p-value calculated using the method specificied in the
pvalue.method parameter.}

\item{pvalue.method}{String. Default is "chisq". Tests association of information added to identifed clusters using this method.}

\item{cex}{Integer. Text size. Default set to 0.8}

\item{sep.outer}{Integer. Default is 0.2.}

\item{border}{String. Color border. Default is NA. Can be set to desired color.}

\item{lwd}{Integer. Default is 0.5. Thickness of border.}
}
\description{
Add a color bar with additional information to the sides of the heatmap.
}
