% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermap.R
\name{plot.gap}
\alias{plot.gap}
\title{plot.gap}
\usage{
\method{plot}{gap}(clust = "col", B = 100, K = 6)
}
\arguments{
\item{clust}{String. Sets direction of clustering, default is "col" (column-wise)}

\item{B}{Integer.}

\item{K}{Integer}
}
\description{
Plots gap statistic
}
