% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermap.R
\name{plot.hmap.key}
\alias{plot.hmap.key}
\title{plot.hmap.key}
\usage{
\method{plot}{hmap.key}()
}
\description{
The function plot.hmap.key() creates a color key, i.e. a graphical overview
of what numerical values the different colors found in the heatmap correspond to.
}
