% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermap.R
\name{plot.hmap}
\alias{plot.hmap}
\title{plot.hmap}
\usage{
\method{plot}{hmap}(X, colorscale = "blue-white-red", xmax = NA,
  col.na = "grey", as.image = T, interpolate = F)
}
\arguments{
\item{X}{matrix}

\item{colorscale}{String. Three (valid R) colors separated by hyphens used to represent the values as colors
in the heatmap. Default is "blue-white-red".}

\item{xmax}{Integer Default is set to maximum value of input matrix X.}

\item{col.na}{String. Default is set to "grey". Specifies color of missing data.}

\item{as.image}{Boolean. Default is set to T}

\item{interpolate}{Boolean. Default is set to F.}

\item{origin}{Integer. Default is set to 0.}
}
\description{
plots heatmaps of a numerical data matrix X.
}
