% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermap.R
\name{plot.init}
\alias{plot.init}
\title{plot.init()}
\usage{
\method{plot}{init}(tree = c(), text = c(), cbar = c(), ckey = T,
  inner = c(), outer = c())
}
\arguments{
\item{tree}{Integer. Adds a dendogram to the plot on the sides specified by the argument. Default is c(2,3),
which adds a dendogram on the left and top sides of the plot.}

\item{text}{Integer. Adds text.}

\item{cbar}{Integer. Adds a color bar.}

\item{ckey}{Boolean. Default is TRUE.}

\item{inner}{Integer vector of 4 elements. Adds (or subtracts) a specific amount of space in the inner margins
on a specified side of the heatmap.}

\item{outer}{Integer vector of 4 elements. Adds (or subtracts) a specific amount of space in the outer margins
on a specified side of the heatmap.}
}
\description{
Initiates a clustermap plot. The purpose is to make
room on the four sides of the heatmap for additional elements such
as cluster dendrograms, color bars with additional info, and labels.
Supply as arguments the sides where you want such elements, with the
coding side=1 (below), side=2 (left), side=3 (top), side=4 (right).
}
\examples{
\dontrun{
plot.init() # Make no room for additional elements
plot.init(tree=c()) # Dendrogram on the left and on the top
plot.init(tree=3, text=4)  # Dendrogram on the top, labels on the right
}
}
