% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermap.R
\name{plot.mean.silhouette}
\alias{plot.mean.silhouette}
\title{plot.mean.silhouette}
\usage{
\method{plot}{mean.silhouette}(clust = "col", K = 10)
}
\arguments{
\item{clust}{String. Sets direction of clustering, default is "col" (column-wise)}

\item{K}{Integer, Default is set to 10.}
}
\description{
Plot mean silhouettes
}
