% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermap.R
\name{plot.silhouette}
\alias{plot.silhouette}
\title{plot.silhouette}
\usage{
\method{plot}{silhouette}(clust = "col", ...)
}
\arguments{
\item{clust}{String. Sets direction of clustering, default is "col" (column-wise)}

\item{order}{String. Default is set to "tree"}
}
\description{
Plot Silhouettes for cluster diagnostics.
}
