% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermap.R
\name{plot.text}
\alias{plot.text}
\title{plot.text}
\usage{
\method{plot}{text}(txt, side = 4, sep.outer = 0, cex = 0.5,
  maxchar = NA)
}
\arguments{
\item{txt}{String. Text to be added to the heatmap.}

\item{side}{Integer. Default is 4 (right side). Defines the side of the heatmap onto which the text will be added.}

\item{sep.outer}{Integer. Default is 0.}

\item{cex}{Integer. Set text size.}

\item{maxchar}{Integer. Default is NA. Truncates the text labels to a maximum number of characters.}
}
\description{
Add labels to the sides of the heatmap.
}
