% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_map.R
\name{purity_cut}
\alias{purity_cut}
\title{purity_cut}
\usage{
purity_cut(hcluster, cutoff = 0.1)
}
\arguments{
\item{hcluster}{A hclust object.}

\item{cutoff}{The edge length cutoff to decide the sub-nodes to merge or not. DEFAULT is 0.1.}
}
\value{
A dataframe of the matching results.
}
\description{
Cut hierarchical clustering dendrogram by edge length and purity of the nodes.
}
