% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recolor.R
\name{recolor_comb}
\alias{recolor_comb}
\title{recolor_comb}
\usage{
recolor_comb(comb_obj, new_group_list, output, comb_delim = "-",
  color = NULL)
}
\arguments{
\item{comb_obj}{A Seurat object for the combined sample. Cells in different samples are labelled by the sample names with the comb_delim.}

\item{new_group_list}{A list of vectors of new group assignment outputted from recolor_s.}

\item{output}{The output directory to save the plot.}

\item{comb_delim}{The delimiter used in the cell names in the combined object to connect sample name and cell name. DEFAULT is '-'.}

\item{color}{A vector of colors used to recolor the new groups. DEFAULT is NULL. Pre defined internal color will be used.}
}
\value{
A vector of new group labels with the cell name as the vector name.
}
\description{
Recolor the combined sample based on the matching results from recolor_s.
}
