% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recolor.R
\name{recolor_s}
\alias{recolor_s}
\title{recolor_s}
\usage{
recolor_s(mapRes_sub, obj, output, color = NULL)
}
\arguments{
\item{mapRes_sub}{A vector of the column named by the sample in the output of cluster_map_by_marker, with the regroup column as the vector name.}

\item{obj}{A Seurat object for the sample.}

\item{output}{The output directory to save the plot.}

\item{color}{A vector of colors used to recolor the new groups. DEFAULT is NULL. Pre defined internal color will be used.}
}
\value{
A vector of new group labels with the cell name as the vector name.
}
\description{
Recolor a single sample based on the matching results from cluster_map_by_marker.
}
