% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separability.R
\name{separability}
\alias{separability}
\title{separability}
\usage{
separability(coord, class_label, k = 5)
}
\arguments{
\item{coord}{A dataframe of the two dimension t-SNE coordinates of each cell in the combined sample.}

\item{class_label}{A vector of sample labels for each cell, with the same order as the tsne_coord.}

\item{k}{K-nearest neighbours used to calculate distance. DEFAULT is 5.}
}
\value{
A single value of separability.
}
\description{
Calculate separability for labeled data. Internal function called by separability_by_group.
}
