% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separability.R
\name{separability_by_group}
\alias{separability_by_group}
\title{separability_by_group}
\usage{
separability_by_group(tsne_coord, group, sample_label, k = 5)
}
\arguments{
\item{tsne_coord}{A dataframe of the two dimension t-SNE coordinates of each cell in the combined sample.}

\item{group}{A vector of group assignment for each cell, with the same order as the tsne_coord.}

\item{sample_label}{A vector of sample labels for each cell, with the same order as the tsne_coord.}

\item{k}{K-nearest neighbours used to calculate distance. DEFAULT is 5.}
}
\value{
A vector of separability for each group.
}
\description{
Calculate separability for each group in a pair of samples. Internal function called by separability_pairwise.
}
