% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separability.R
\name{separability_pairwise}
\alias{separability_pairwise}
\title{separability_pairwise}
\usage{
separability_pairwise(tsne_coord, group, sample_label, k = 5)
}
\arguments{
\item{tsne_coord}{A dataframe of the two dimension t-SNE coordinates of each cell in the combined sample.}

\item{group}{A vector of group assignment for each cell, with the same order as the tsne_coord.}

\item{sample_label}{A vector of sample labels for each cell, with the same order as the tsne_coord.}

\item{k}{K-nearest neighbours used to calculate distance. DEFAULT is 5.}
}
\value{
A matrix of separability for each sample pair (column) and each group (row).
}
\description{
Calculate separability for every sample pair. The higher the more separable.
}
