% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermap.R
\name{set.color}
\alias{set.color}
\title{set.color}
\usage{
set.color(x, type = "discrete", label, color, na.color, clust)
}
\arguments{
\item{x}{matrix}

\item{type}{String. Default is "discrete".}

\item{label}{String. Default is "".}

\item{color}{String. Default is "".}

\item{na.color}{String. Set colors to missing data. Default is "grey".}

\item{clust}{String. Default is "col". Specifies column-wise ("col") or row-rise clustering ("row").}
}
\description{
Converts numerical information into colors
}
