% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermap.R
\name{subclust}
\alias{subclust}
\title{subclust}
\usage{
subclust(k = NA, clust = "col", method = "gap", B = 50, min.size = 5,
  max.level = 3)
}
\arguments{
\item{k}{Integer. Default is NA. Desired number of clusters. If default, k is estimated automatically.}

\item{clust}{String. Identify clusters column-wise if set to "col" (default), or row-wise if set to "row".}

\item{method}{String. Default is "gap". If k is NA, the number of clusters that automatically be estimated
by one of two algorithms: gap (default) (Tibshirani et al. (2001), J Roy Statist Soc B, 63: 411-423) or PART
(Nilsen et al (2012), Stat Appl Genet Mol Biol, 12: 637-652).}

\item{B}{Integer. Number of permutations. Used if method set to 'PART'.}

\item{min.size}{Integer. Least number of element in a single cluster. Used if method set to 'PART'.}

\item{max.level}{Integer. The level in the dendogram counting from the root to stop searching for clusters.}
}
\description{
The function identifies subclusters in a dendrogram.
The effect is that later calls to draw.tree will show the subclusters
in different colors. You can either provide the desired number of
clusters using the argument k, or you can set k=NA to automatically
estimate the number of cluster with one of two algorithms:
gap (Tibshirani et al. (2001), J Roy Statist Soc B, 63: 411-423) or
PART  (Nilsen et al (2012), Stat Appl Genet Mol Biol, 12: 637-652).
}
