% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustermap.R
\docType{package}
\name{clustermap}
\alias{clustermap}
\alias{clustermap-package}
\title{clustermap}
\description{
clustermap is a package for hierarchical clustering and heatmap visualisation of numerical
data matrices.
}
\author{
Ole Christian Lingjærde & Chloé B. Steen
}
